requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
requests>=2.31.0
pynput>=1.7.0
